\name{checkerboard}

\alias{checkerboard}

\title{ Checkerboard statistic }

\description{
  Calculates the checkerboard statistic of species co-occurrence
}

\usage{
checkerboard(x)
}

\arguments{
  \item{x}{ Community data matrix }
}

\value{
  Return an object of class dist containing checkerboard index of co-occurrence for all species pairs in x.
}
\references{ ~put references to the literature/web site here ~ }
\author{ Steve Kembel <skembel@berkeley.edu> }
\keyword{univar}
