% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylobeta.R
\name{comdist}
\alias{comdist}
\title{Calculates inter-community mean pairwise distance}
\usage{
comdist(comm, dis, abundance.weighted = FALSE, Rcpp = TRUE, threads = 1)
}
\arguments{
\item{comm}{Community data matrix}

\item{dis}{Interspecific distance matrix}

\item{abundance.weighted}{Should mean pairwise distances separating species
in two communities be weighted by species abundances? (default = FALSE)}

\item{threads}{is an interger of number of threads for parallel processing}
}
\value{
Distance object of MPD values separating each pair of communities.
}
\description{
Calculates MPD (mean pairwise distance) separating taxa in two communities,
a measure of phylogenetic beta diversity
}
\details{
This function calculates a measure of phylogenetic beta diversity: the
expected phylogenetic distance separating two individuals or taxa drawn
randomly from different communities.
}
\examples{

data(phylocom)
comdist(phylocom$sample, cophenetic(phylocom$phylo), abundance.weighted=TRUE)
}
\references{
C.O. Webb, D.D. Ackerly, and S.W. Kembel. 2008. Phylocom:
software for the analysis of phylogenetic community structure and trait
evolution. Bioinformatics 18:2098-2100.
}
\seealso{
\code{\link{mpd}}, \code{\link{ses.mpd}}
}
\author{
Steven Kembel \href{mailto:steve.kembel@gmail.com}{steve.kembel@gmail.com}
}
\keyword{univar}
