\name{contrast.cor.P}
\alias{contrast.cor.P}
\title{P-value for PIC correlation with corrected df}
\description{
  P-value for PIC correlation with corrected df
}
\usage{
contrast.cor.P(r, df)
}
\arguments{
  \item{r}{ Correlation coefficient }
  \item{df}{ Degrees of freedom }
}
\author{ Steve Kembel <skembel@berkeley.edu> }
\keyword{univar}
