% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample.R
\name{datasetsVerification}
\alias{datasetsVerification}
\title{datasetsVerification}
\usage{
datasetsVerification(dataset)
}
\arguments{
\item{dataset}{the dataset to verify}
}
\description{
This fonction test an dataset for commons errors
}
\details{
This fonction look for if it :
\itemize{
is an data.frame
is empty
contain NA
contain characters
}
}
