\name{freq.dist}
\alias{freq.dist}

\title{ Frequency distance }
\description{
  Dissimilarity of species occurrence frequencies as a distance object
}
\usage{
freq.dist(sppFreq, metric = "rank")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sppFreq}{ Vector of species frequencies }
  \item{metric}{ freq or rank }
}
\value{
  A dist object
}
\author{ Steve Kembel <skembel@berkeley.edu> }
