\name{get.nodes}
\Rdversion{1.1}
\alias{get.nodes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
gets nodes along the evolutionary path to a given species
}
\description{
Get a vector of the nodes that are along the path from the root to spp x.  Used in other functions.
}
\usage{
get.nodes(tree, spp)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{
APP tree
}
  \item{spp}{
specific species; same format as which.edge in the APE library
}
}


\author{
Karen Magnuson-Ford 
Will Cornwell
Arne Mooers
Mark Vellend
}



