\name{internal2tips}
\alias{internal2tips}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Tips descended from a node }
\description{
  Tips descended from a node
}
\usage{
internal2tips(x, int.node, return.names = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ phylo object }
  \item{int.node}{ internal node number }
  \item{return.names}{ TRUE or FALSE }
}
\value{
  Vector of tips descended from a node
}

\author{ David Ackerly <dackerly@berkeley.edu> }
\keyword{manip}
