% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.checking.R
\name{match.phylo.comm}
\alias{match.phylo.comm}
\title{Match taxa in phylogeny and data}
\usage{
match.phylo.comm(phy, comm)
}
\arguments{
\item{phy}{A phylogeny object of class phylo}

\item{comm}{Community data matrix
match.phylo.comm(phy = phylocom$phylo,comm = phylocom$sample)}
}
\description{
Match taxa in phylogeny and data
}
