% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylocom.R
\name{matrix2sample}
\alias{matrix2sample}
\title{Convert community data matrix to Phylocom sample}
\usage{
matrix2sample(z)
}
\arguments{
\item{z}{Community data matrix}
}
\value{
Phylocom database-format community sample
}
\description{
Converts a community data matrix to a Phylocom database-format community
sample
}
\examples{

data(phylocom)
matrix2sample(phylocom$sample)

}
\references{
Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom:
software for the analysis of phylogenetic community structure and trait
evolution. Version 4.0.1. \url{http://www.phylodiversity.net/phylocom/}.
}
\author{
Steven Kembel \href{mailto:steve.kembel@gmail.com}{steve.kembel@gmail.com} and Cam Webb
\href{mailto:cwebb@oeb.harvard.edu}{cwebb@oeb.harvard.edu}
}
\keyword{manip}
