% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodiversity.R
\name{mntd}
\alias{mntd}
\alias{mnnd}
\title{Mean nearest taxon distance}
\usage{
mntd(samp, dis, abundance.weighted = FALSE)
}
\arguments{
\item{samp}{Community data matrix}

\item{dis}{Interspecific distance matrix}

\item{abundance.weighted}{Should mean nearest taxon distances for each
species be weighted by species abundance? (default = FALSE)}
}
\value{
Vector of MNTD values for each community.
}
\description{
Calculates MNTD (mean nearest taxon distance) for taxa in a community
}
\details{
This metric has also been referred to as MNND (mean nearest neighdata:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAWElEQVR42mNgGPTAxsZmJsVqQApgmGw1yApwKcQiT7phRBuCzzCSDSHGMKINIeDNmWQlA2IigKJwIssQkHdINgxfmBBtGDEBS3KCxBc7pMQgMYE5c/AXPwAwSX4lV3pTWwAAAABJRU5ErkJggg==bour
distance), and the function was named \code{mnnd} in picante versions <
0.7.
}
\examples{
data(phylocom)
mntd(phylocom$sample, cophenetic(phylocom$phylo), abundance.weighted=TRUE)

}
\references{
Webb, C., D. Ackerly, M. McPeek, and M. Donoghue. 2002.
Phylogenies and community ecology. Annual Review of Ecology and Systematics
33:475-505.
}
\seealso{
\code{\link{ses.mntd}}
}
\author{
Steven Kembel \href{mailto:steve.kembel@gmail.com}{steve.kembel@gmail.com}
}
\keyword{univar}
