\name{node.age}
\alias{node.age}

\title{ Node ages }
\description{
  Node ages based on branch lengths
}
\usage{
node.age(phy)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{phy}{ phylo object }
}

\value{
  Phylo object with phylo\$ages vector of node ages corresponding to phylo\$edge
}

\author{ David Ackerly <dackerly@berkeley.edu> }
\keyword{misc}
