\name{node.desc}
\Rdversion{1.1}
\alias{node.desc}
\title{
Descendants of a particular node
}
\description{
Finds the descendants from a particular node
}
\usage{
node.desc(tr, node)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tr}{
an object of class phylo
}
  \item{node}{
node number
}
}


\note{
utility function.  modified from node.sons from  Walton Green, Gene Hunt, Scott Wing 
This function could easily be hidden for simplicity's sake.

}


