\name{node.number}
\Rdversion{1.1}
\alias{node.number}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get number of internal nodes for each species
}
\description{
get the number of internal nodes for each spp
}
\usage{
node.number(tree)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tree}{
APE phylogeny
}
}


\author{
Will Cornwell
}
\note{
function for internal use
}


