% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.indexes.R
\name{NTI}
\alias{NTI}
\title{nearest taxon index (NTI)}
\usage{
NTI(
  samp,
  dis,
  null.model = c("taxa.labels", "richness", "frequency", "sample.pool", "phylogeny.pool",
    "independentswap", "trialswap"),
  abundance.weighted = FALSE,
  runs = 999,
  iterations = 1000
)
}
\arguments{
\item{samp}{Community data matrix}

\item{dis}{Distance matrix (generally a phylogenetic distance matrix)}

\item{null.model}{Null model to use (see Details section for description)}

\item{abundance.weighted}{Should mean nearest taxon distances for each
species be weighted by species abundance? (default = FALSE)}

\item{runs}{Number of randomizations}

\item{iterations}{Number of iterations to use for each randomization (for}
}
\description{
calculates the mean nearest phylogenetic neighbor among
the individuals in a community
}
\section{Functions}{
\itemize{
\item \code{NTI()}: 

}}
\references{
{Chai, Y., Yue, M., Liu, X. et al. Patterns of taxonomic,
phylogenetic diversity during a long-term succession of forest
on the Loess Plateau, China: insights into assembly process.
Sci Rep 6, 27087 (2016). https://doi.org/10.1038/srep27087}
}
