\name{phylo2phylog}
\alias{phylo2phylog}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Convert ape phylo to ade4 phylog }
\description{
  Convert ape phylo to ade4 phylog
}
\usage{
phylo2phylog(phy, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{phy}{ phylo object }
  \item{\dots}{ Additional arguments to newick2phylog }
}

\value{
  phylog object
}

\author{ Steve Kembel <skembel@berkeley.edu> }
\keyword{manip}
