% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosor.R
\name{phylosor}
\alias{phylosor}
\title{Phylogenetic index of beta-diversity PhyloSor}
\usage{
phylosor(samp, tree)
}
\arguments{
\item{samp}{Community data matrix}

\item{tree}{Object of class phylo - a rooted phylogeny}
}
\value{
A distance object of the PhyloSor index of similarity between
communities, the fraction of PD (branch-length) shared between two samples
}
\description{
Fraction of branch-length shared between two communities
}
\note{
The root of the supplied tree is included in calculations of
PhyloSor. The supplied tree must be rooted. Single-species samples will be
assigned a PD value equal to the distance from the root to the present.
}
\section{Warning }{
 The phylosor of all samples will include the branch
length connecting taxa in those samples and the root of the supplied tree.
The root of the supplied tree may not be spanned by any taxa in the sample.
If you want the root of your tree to correspond to the most recent ancestor
of the taxa actually present in your sample, you should prune the tree
before running \code{phylosor}:

\code{prunedTree <- prune.sample(sample,tree)}
}

\examples{

data(phylocom)
phylosor(phylocom$sample, phylocom$phylo)
}
\references{
Bryant, J.B., Lamanna, C., Morlon, H., Kerkhoff, A.J., Enquist,
B.J., Green, J.L. 2008. Microbes on mountainsides: Contrasting elevational
patterns of bacterial and plant diversity. Proceedings of the National
Academy of Sciences 105 Supplement 1: 11505-11511
}
\seealso{
\code{\link{phylosor.rnd}}, \code{\link{pd}}
}
\author{
Helene Morlon \href{mailto:morlon.helene@gmail.com}{morlon.helene@gmail.com} and Steven Kembel
\href{mailto:steve.kembel@gmail.com}{steve.kembel@gmail.com}
}
\keyword{univar}
