\name{picante-package}
\alias{picante-package}
\alias{picante}
\docType{package}
\title{
Picante: Phylocom Integration, Community Analyses, Null-models, Traits, and Evolution in R
}
\description{
Collection of functions for integrating ecology and evolutionary biology
}
\details{
\tabular{ll}{
Package: \tab picante\cr
Type: \tab Package\cr
Version: \tab 0.5-0\cr
Date: \tab 2009-1-22\cr
License: \tab GPL-2\cr
}
}
\author{
Steven Kembel <skembel@uoregon.edu>, David Ackerly <dackerly@berkeley.edu>, Simon Blomberg <s.blomberg1@uq.edu.au>, Peter Cowan <pdc@berkeley.edu>, Matthew Helmus <mrhelmus@wisc.edu>, Helene Morlon  <morlon.helene@gmail.com>, Campbell Webb <cwebb@oeb.harvard.edu>

Maintainer: Steven Kembel <skembel@uoregon.edu>
}
\keyword{package}
