% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample.R
\name{presenceAbsenceConverter}
\alias{presenceAbsenceConverter}
\title{presence-absence converter}
\usage{
presenceAbsenceConverter(dataset)
}
\arguments{
\item{dataset}{an dataframe to convert. The values in the dataset have to be
equal or superior to zero an to be an integer}
}
\value{
an dataset of only 0,1 values
}
\description{
Convert an data.frame to an dataset containing only 0,1 values
}
\examples{
data(phylocom)
presenceAbsenceConverter(phylocom$sample)
}
