% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{prune.sample}
\alias{prune.sample}
\alias{prune.missing}
\title{Prune tree to match community data or trait data}
\usage{
prune.sample(samp, phylo)

prune.missing(x, phylo)
}
\arguments{
\item{samp}{Community data matrix}

\item{phylo}{phylo object}

\item{x}{Vector of trait data}
}
\value{
Returns a pruned phylo object
}
\description{
Prune a phylogenetic tree to include only species present in a community
data set or with non-missing trait data
}
\section{Functions}{
\itemize{
\item \code{prune.missing()}: Prune a phylogenetic tree to include non-missing
trait data

}}
\author{
Steven Kembel \href{mailto:steve.kembel@gmail.com}{steve.kembel@gmail.com}
}
\keyword{manip}
