\name{randomizeSampleKeepSampRichness}
\alias{randomizeSampleKeepSampRichness}

\title{ Randomize community matrix maintaining sample richness }
\description{
  Randomize community matrix maintaining sample richness
}
\usage{
randomizeSampleKeepSampRichness(samp)
}
\arguments{
  \item{samp}{ Community data matrix }
}
\value{
 Randomized community data matrix
}
\references{ Gotelli. 2000. Ecology. }
\author{ Steve Kembel <skembel@berkeley.edu> }
