\name{randomizeSampleKeepSppFreq}
\alias{randomizeSampleKeepSppFreq}
\title{ Randomize community matrix maintaining species frequencies }
\description{
  Randomize community matrix maintaining species frequencies
}
\usage{
randomizeSampleKeepSppFreq(samp)
}
\arguments{
  \item{samp}{ Community data matrix }
}
\value{
 Randomized community data matrix
}
\references{ Gotelli. 2000. Ecology. }
\author{ Steve Kembel <skembel@berkeley.edu> }
