% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylocom.R
\name{readsample}
\alias{readsample}
\title{Read Phylocom sample}
\usage{
readsample(filename = "")
}
\arguments{
\item{filename}{Phylocom sample file path}
}
\value{
Community data matrix
}
\description{
Reads a Phylocom sample file and converts to a community data matrix
}
\references{
Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom:
software for the analysis of phylogenetic community structure and trait
evolution. Version 4.0.1. \url{http://www.phylodiversity.net/phylocom/}.
}
\author{
Steven Kembel \if{html}{\out{<skembel>}} and Cam Webb \href{mailto:cwebb@oeb.harvard.edu}{cwebb@oeb.harvard.edu}
}
\keyword{IO}
