\name{reflect.contrasts}
\alias{reflect.contrasts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Reflect nodal contrast values about the origin }
\description{
  Reflect nodal contrast values about the origin for correlation calculation.
}
\usage{
reflect.contrasts(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Data frame of nodal PIC values }
}

\value{
  Data frame of nodal PIC values reflected about the origin
}

\author{ Steve Kembel <skembel@berkeley.edu> }
