\name{resample}
\alias{resample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Improved sample }
\description{
  Utility function cribbed from \code{\link{sample}} documentation.
}
\usage{
resample(x, size, ...)
}
\arguments{
  \item{x}{ Object to be sampled from}
  \item{size}{ Number of samples to take}
  \item{...}{ Parameters to be passed on}
}


\seealso{ \code{\link{sample}} }
\keyword{manip}
