\name{roij}
\alias{roij}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Frequency-insensitive measure of species co-occurrence }
\description{
  Frequency-insensitive measure of species co-occurrence, based on modified checkerboard score.
}
\usage{
roij(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Community data matrix }
}

\value{
  Dist object of species co-occurrence.
}
\references{ Hardy. pers. comm. }
\author{ Steve Kembel <skembel@berkeley.edu> }
\keyword{univar}
