% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample.R
\name{rsample}
\alias{rsample}
\title{rsample}
\usage{
rsample(maximum, tree, all_NA = FALSE, column_with_NA = FALSE)
}
\arguments{
\item{maximum}{the maximum value inside an cell}

\item{tree}{A phylo tree object}

\item{all_NA}{if TRUE, all cells will be NA}

\item{column_with_NA}{if insered, the cell in the column will all be NA}
}
\description{
This fonction generate an community data matrix
}
\author{
Kerby Elpenord
}
\keyword{community}
\keyword{generate,}
\keyword{sample}
