\name{sample2matrix}
\alias{sample2matrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Convert Phylocom sample to community data matrix }
\description{
  Convert Phylocom sample to community data matrix
}
\usage{
sample2matrix(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Phylocom sample formatted data frame }
}

\references{ Webb et al. 2007. Phylocom. }
\author{ Steve Kembel <skembel@berkeley.edu> and Cam Webb <cwebb@oeb.harvard.edu> }
\keyword{IO}
