\name{shared.history}
\alias{shared.history}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Shared history of a tree }
\description{
  Shared history of a tree
}
\usage{
shared.history(t)
}

\arguments{
  \item{t}{ phylo object }
}

\value{
  Numeric shared history estimate
}

\author{ David Ackerly <dackerly@berkeley.edu> }
\keyword{univar}
