\name{sortColumns}
\alias{sortColumns}

\title{ Sort columns of a data frame or matrix}
\description{
  Utility function to sort columns of a data frame or matrix
}
\usage{
sortColumns(x)
}

\arguments{
  \item{x}{ Data frame or matrix }
}

\value{
  Sorted data frame or matrix
}

\author{ Steve Kembel <skembel@berkeley.edu> }
