\name{sortRows}
\alias{sortRows}
\title{ Sort rows of a data frame or matrix}
\description{
  Utility function to sort rows of a data frame or matrix
}
\usage{
sortRows(x)
}

\arguments{
  \item{x}{ Data frame or matrix }
}

\value{
  Sorted data frame or matrix
}

\author{ Steve Kembel <skembel@berkeley.edu> }
\keyword{manip}
