\name{sppFreq}
\alias{sppFreq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Calculate species occurrence frequencies }
\description{
  Calculate species occurrence frequencies 
}
\usage{
sppFreq(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Community data matrix }
}

\author{ Steve Kembel <skembel@berkeley.edu> }

\seealso{ \code{\link{decostand}} }
\keyword{univar}
