\name{phylocom.stack}
\alias{phylocom.stack}

\title{ Stack a database-format community data frame into a community data matrix }
\description{
  Stack a database-format community data frame into a community data matrix. Mostly utility function for converting Phylocom sample format data frames into community data matrices.
}
\usage{
phylocom.stack(z)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{ Phylocom format sample data frame }
}

\author{ Cam Webb <cwebb@oeb.harvard.edu> }
\keyword{manip}
