% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylocom.R
\name{writesample}
\alias{writesample}
\title{Write a Phylocom community sample file}
\usage{
writesample(community, filename = "")
}
\arguments{
\item{community}{Community data matrix}

\item{filename}{Filename path}
}
\description{
Write a community data matrix to a Phylocom community sample file
}
\references{
Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom:
software for the analysis of phylogenetic community structure and trait
evolution. Version 4.0.1. \url{http://www.phylodiversity.net/phylocom/}.
}
\author{
Steven Kembel \href{mailto:steve.kembel@gmail.com}{steve.kembel@gmail.com} and Cam Webb
\href{mailto:cwebb@oeb.harvard.edu}{cwebb@oeb.harvard.edu}
}
\keyword{file}
