% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylocom.R
\name{writetraits}
\alias{writetraits}
\title{Write a Phylocom traits formatted file}
\usage{
writetraits(trt, file = "", bin = NULL, sigd = 3)
}
\arguments{
\item{trt}{Data frame containing trait data}

\item{file}{Filename path}

\item{bin}{Vector index of trait columns to be treated as binary}

\item{sigd}{Significant digits for output}
}
\description{
Write a Phylocom traits formatted file
}
\references{
Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom:
software for the analysis of phylogenetic community structure and trait
evolution. Version 4.0.1. \url{http://www.phylodiversity.net/phylocom/}.
}
\author{
David Ackerly \href{mailto:dackerly@berkeley.edu}{dackerly@berkeley.edu} and Steven Kembel
\href{mailto:steve.kembel@gmail.com}{steve.kembel@gmail.com}
}
\keyword{file}
