% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.table.R
\name{cor.table}
\alias{cor.table}
\title{Table of correlations and P-values}
\usage{
cor.table(x, cor.method = c("pearson","spearman"),
cor.type=c("standard","contrast"))
}
\arguments{
\item{x}{Data frame of data points or contrasts at nodes}

\item{cor.method}{Correlation method}

\item{cor.type}{Are data standard or independent contrast values?}
}
\description{
Table of correlations with associated P-values and df,
can be used with regular or independent contrast data
}
\examples{
example_1 <- data.frame(
X = c(4.09434, 3.61092, 2.37024, 2.02815, -1.46968),
Y = c(4.74493, 3.33220, 3.36730, 2.89037, 2.30259))
cor.table(x          = example_1, 
         cor.method = "pearson",
         cor.type   = "contrast")
}
\references{
Garland, T., Jr., P. H. Harvey, and A. R. Ives. 1992.
Procedures for the analysis of comparative data using phylogenetically
independent contrasts. Systematic Biology 41:18-32.
Table of correlations and P-values
Table of correlations with associated P-values and df, can be used with
regular or independent contrast data
}
\keyword{univar}
