% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosignal.R
\name{Kcalc}
\alias{Kcalc}
\title{K statistic of phylogenetic signal}
\usage{
Kcalc(x, phy, checkdata = TRUE)
}
\arguments{
\item{x}{Vector or data.frame of trait data (in phylo$tip.label order)}

\item{phy}{phylo object}

\item{checkdata}{Check for match between trait and phylogeny taxa labels
using \code{\link{match.phylo.data}}? (default=TRUE)}
}
\value{
\item{K }{K statistic}
}
\description{
Calculates K statistic of phylogenetic signal
}
\examples{

\donttest{
randtree <- rcoal(20)
randtraits <- rTraitCont(randtree)
Kcalc(randtraits[randtree$tip.label],randtree)
}

}
\references{
Blomberg, S. P., and T. Garland, Jr. 2002. Tempo and mode in
evolution: phylogenetic inertia, adaptation and comparative methods.
Journal of Evolutionary Biology 15:899-910.

Blomberg, S. P., T. Garland, Jr., and A. R. Ives. 2003. Testing for
phylogenetic signal in comparative data: behavioral traits are more labile.
Evolution 57:717-745.
}
\seealso{
\code{\link{phylosignal}}
}
\author{
Simon Blomberg \href{mailto:s.blomberg1@uq.edu.au}{s.blomberg1@uq.edu.au} and David Ackerly
\href{mailto:dackerly@berkeley.edu}{dackerly@berkeley.edu}
}
\keyword{univar}
