\name{matrix2sample}
\alias{matrix2sample}

\title{ Convert community data matrix to Phylocom sample }
\description{
  Converts a community data matrix to a Phylocom database-format community sample
}
\usage{
matrix2sample(z)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{ Community data matrix }
}

\value{
  Phylocom database-format community sample
}
\references{ Webb et al. 2007. Phylocom. }
\author{ Steve Kembel <skembel@berkeley.edu> and Cam Webb <cwebb@oeb.harvard.edu> }
