% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodiversity.R
\name{mpd}
\alias{mpd}
\title{Mean pairwise distance}
\usage{
mpd(samp, dis, abundance.weighted = FALSE)
}
\arguments{
\item{samp}{Community data matrix}

\item{dis}{Interspecific distance matrix}

\item{abundance.weighted}{Should mean pairwise distances be weighted by
species abundance? (default = FALSE)}
}
\value{
Vector of MPD values for each community
}
\description{
Calculates mean pairwise distance separating taxa in a community
}
\examples{
data(phylocom)
mpd(phylocom$sample, cophenetic(phylocom$phylo), abundance.weighted=TRUE)

}
\references{
Webb, C., D. Ackerly, M. McPeek, and M. Donoghue. 2002.
Phylogenies and community ecology. Annual Review of Ecology and Systematics
33:475-505.
}
\seealso{
\code{\link{ses.mpd}}
}
\author{
Steven Kembel \href{mailto:steve.kembel@gmail.com}{steve.kembel@gmail.com}
}
\keyword{univar}
