\name{phylocom}
\alias{phylocom}
\docType{data}
\title{ Phylocom default data }
\description{
  Phylogeny, community and trait data from the Phylocom 4.0 distribution
}
\usage{data(phylocom)}
\format{
    A list with three elements:
    \itemize{
      \item{ phylo }{ Phylogenetic tree (an object of class phylo) }
      \item{ sample }{ Community data (a data.frame with samples in rows and species in columns }
      \item{ traits }{ Trait data (a data.frame with species in rows and traits in columns }
      }
}

\source{
Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom: software for the analysis of phylogenetic community structure and trait evolution. Version 4.0.1.
\url{http://www.phylodiversity.net/phylocom/}.
}

\keyword{datasets}
