\name{picante-package}
\alias{picante-package}
\alias{picante}
\docType{package}
\title{
Picante: Phylocom Integration, Community Analyses, Null-models, Traits, and Evolution in R
}
\description{
Collection of functions for integrating ecology and evolutionary biology.
}
\details{
\tabular{ll}{
Package: \tab picante\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2007-12-05\cr
License: \tab GPL v2\cr
}
More details forthcoming
}
\author{
Steve Kembel <skembel@berkeley.edu>, David Ackerly <dackerly@berkeley.edu>, Simon Blomberg <s.blomberg1@uq.edu.au>, Cam Webb <cwebb@oeb.harvard.edu>

Maintainer: Steve Kembel <skembel@berkeley.edu>
}
