\name{picante-package}
\alias{picante-package}
\alias{picante}
\docType{package}
\title{
picante: Integrating Phylogenies and Ecology
}
\description{
Phylocom integration, community analyses, null-models, traits and evolution. Ecophylogenetic approaches including measures of community phylogenetic and trait diversity. Phylogenetic signal. Estimation of trait values for unobserved taxa. Null models for community and phylogeny randomizations. Utility functions for data input/output and phylogeny plotting.
}
\details{
\tabular{ll}{
Package: \tab picante\cr
Type: \tab Package\cr
Version: \tab 1.7\cr
Date: \tab 2018-05-01\cr
License: \tab GPL-2\cr
}
}
\author{
Author: Steven W. Kembel <steve.kembel@gmail.com>, David D. Ackerly <dackerly@berkeley.edu>, Simon P. Blomberg <s.blomberg1@uq.edu.au>, Will K. Cornwell <cornwell@zoology.ubc.ca>, Peter D. Cowan <pdc@berkeley.edu>, Matthew R. Helmus <mrhelmus@wisc.edu>, Helene Morlon <morlon.helene@gmail.com>, Campbell O. Webb <cwebb@oeb.harvard.edu>
Maintainer: Steven W. Kembel <steve.kembel@gmail.com>
}
\keyword{package}