\name{picante-package}
\alias{picante-package}
\alias{picante}
\docType{package}
\title{
picante: Integrating Phylogenies and Ecology
}
\description{
Functions for phylocom integration, community analyses, null-models, traits and evolution. Implements numerous ecophylogenetic approaches including measures of community phylogenetic and trait diversity, phylogenetic signal, estimation of trait values for unobserved taxa, null models for community and phylogeny randomizations, and utility functions for data input/output and phylogeny plotting. A full description of package functionality and methods are provided by Kembel et al. (2010) <doi:10.1093/bioinformatics/btq166>.
}
\details{
\tabular{ll}{
Package: \tab picante\cr
Type: \tab Package\cr
Version: \tab 1.8.1\cr
Date: \tab 2020-02-12\cr
License: \tab GPL-2\cr
}
}
\author{
Author: Steven W. Kembel <steve.kembel@gmail.com>, David D. Ackerly <dackerly@berkeley.edu>, Simon P. Blomberg <s.blomberg1@uq.edu.au>, Will K. Cornwell <cornwell@zoology.ubc.ca>, Peter D. Cowan <pdc@berkeley.edu>, Matthew R. Helmus <mrhelmus@wisc.edu>, Helene Morlon <morlon.helene@gmail.com>, Campbell O. Webb <cwebb@oeb.harvard.edu>
Maintainer: Steven W. Kembel <steve.kembel@gmail.com>
}
\keyword{package}