\name{picante-package}
\alias{picante-package}
\alias{picante}
\docType{package}
\title{
picante: R tools for integrating phylogenies and ecology
}
\description{
R tools for integrating phylogenies and ecology
}
\details{
\tabular{ll}{
Package: \tab picante\cr
Type: \tab Package\cr
Version: \tab 1.6-2\cr
Date: \tab 2014-03-05\cr
License: \tab GPL-2\cr
}
}
\author{
Author: Steven W. Kembel <steve.kembel@gmail.com>, David D. Ackerly <dackerly@berkeley.edu>, Simon P. Blomberg <s.blomberg1@uq.edu.au>, Will K. Cornwell <cornwell@zoology.ubc.ca>, Peter D. Cowan <pdc@berkeley.edu>, Matthew R. Helmus <mrhelmus@wisc.edu>, Helene Morlon <morlon.helene@gmail.com>, Campbell O. Webb <cwebb@oeb.harvard.edu>
Maintainer: Steven W. Kembel <steve.kembel@gmail.com>
}
\keyword{package}
