\name{picante-package}
\alias{picante-package}
\alias{picante}
\docType{package}
\title{
picante: R tools for integrating phylogenies and ecology
}
\description{
R tools for integrating phylogenies and ecology
}
\details{
\tabular{ll}{
Package: \tab picante\cr
Type: \tab Package\cr
Version: \tab 0.7-2\cr
Date: \tab 2009-10-27\cr
License: \tab GPL-2\cr
}
}
\author{
Steven Kembel <skembel@uoregon.edu>, David Ackerly <dackerly@berkeley.edu>, Simon Blomberg <s.blomberg1@uq.edu.au>, Peter Cowan <pdc@berkeley.edu>, Matthew Helmus <mrhelmus@wisc.edu>, Helene Morlon  <morlon.helene@gmail.com>, Campbell Webb <cwebb@oeb.harvard.edu>

Maintainer: Steven Kembel <skembel@uoregon.edu>
}
\keyword{package}
