\name{readsample}
\alias{readsample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Read Phylocom sample }
\description{
  Reads a Phylocom sample file and converts to a community data matrix
}
\usage{
readsample(filename = "")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{filename}{ Phylocom sample file path }
}

\value{
  Community data matrix
}
\references{ Webb et al. 2007. Phylocom. }
\author{ Steve Kembel <skembel> and Cam Webb <cwebb@oeb.harvard.edu> }
\keyword{IO}
