\name{readsample}
\alias{readsample}

\title{ Read Phylocom sample }
\description{
  Reads a Phylocom sample file and converts to a community data matrix
}
\usage{
readsample(filename = "")
}

\arguments{
  \item{filename}{ Phylocom sample file path }
}

\value{
  Community data matrix
}
\references{   Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom: software for the analysis of phylogenetic community structure and trait evolution. Version 4.0.1. \url{http://www.phylodiversity.net/phylocom/}. }
\author{ Steven Kembel <skembel> and Cam Webb <cwebb@oeb.harvard.edu> }
\keyword{IO}
