% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylocom.R
\name{sample2matrix}
\alias{sample2matrix}
\title{Convert Phylocom sample to community data matrix}
\usage{
sample2matrix(x)
}
\arguments{
\item{x}{Phylocom sample formatted data frame, a data frame with three\tabular{lll}{
   Community name \tab Species abundance \tab Species name \cr
   0 \tab 0 \tab 0 \cr
}}
}
\description{
Convert a Phylocom database-format sample to community data matrix.
}
\references{
Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom:
software for the analysis of phylogenetic community structure and trait
evolution. Version 4.0.1. \url{http://www.phylodiversity.net/phylocom/}.
}
\author{
Steven Kembel \href{mailto:steve.kembel@gmail.com}{steve.kembel@gmail.com} and Cam Webb
\href{mailto:cwebb@oeb.harvard.edu}{cwebb@oeb.harvard.edu}
}
\keyword{IO}
