\name{writesample}
\alias{writesample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Write a Phylocom community sample file }
\description{
  Write a Phylocom community sample file
}
\usage{
writesample(community, filename = "")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{community}{ Community data matrix }
  \item{filename}{ Filename path }
}
\references{ Webb et al. 2007. Phylocom. }
\author{ Steve Kembel <skembel@berkeley.edu> and Cam Webb <cwebb@oeb.harvard.edu> }
