\name{writesample}
\alias{writesample}

\title{ Write a Phylocom community sample file }
\description{
  Write a community data matrix to a Phylocom community sample file
}
\usage{
writesample(community, filename = "")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{community}{ Community data matrix }
  \item{filename}{ Filename path }
}
\references{ Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom: software for the analysis of phylogenetic community structure and trait evolution. Version 4.0.1. \url{http://www.phylodiversity.net/phylocom/}. }
\author{ Steve Kembel <skembel@berkeley.edu> and Cam Webb <cwebb@oeb.harvard.edu> }
\keyword{file}
